simple-rag-bot/requirements.txt
chromadb==0.4.22
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
sentence-transformers==2.5.1
urllib3==1.26.15
